replace Input.js with MonacoInput.js
