+++ packages/app/src/app/components/sandbox/Preview/DevTools/Console/Input.js
@@ -1,95 +1,188 @@
+/* @flow */
 import React from 'react';
 import styled from 'styled-components';
 
 import ChevronRight from 'react-icons/lib/md/chevron-right';
 import theme from 'common/theme';
 
+import MonacoEditor from 'app/components/sandbox/CodeEditor/monaco/MonacoReactComponent';
+
 import { IconContainer } from './styles';
 
+const CONSOLE_INPUT_TOP_PADDING = 6;
+const CONSOLE_INPUT_BOTTOM_PADDING = 6;
+const CONSOLE_INPUT_PADDING =
+  CONSOLE_INPUT_TOP_PADDING + CONSOLE_INPUT_BOTTOM_PADDING;
+const CONSOLE_INPUT_LINE_HEIGHT = 20;
+const CONSOLE_INPUT_MAX_HEIGHT = 110;
+
 const Container = styled.div`
+  flex-shrink: 0;
   position: relative;
-  height: 2rem;
+  height: ${props => props.height}px;
   min-height: 2rem;
   width: 100%;
   background-color: ${props => props.theme.background.darken(0.3)};
-
   display: flex;
-  align-items: center;
-`;
+  align-items: flex-start;
 
-const Input = styled.input`
-  position: relative;
-  height: 1.5rem;
-  width: 100%;
-  background-color: ${props => props.theme.background.darken(0.3)};
-  border: none;
-  outline: none;
-  color: rgba(255, 255, 255, 0.8);
-  font-family: Menlo, monospace;
-  font-size: 13px;
+  .monaco-editor-background {
+    background-color: ${props => props.theme.background.darken(0.3)};
+  }
 `;
 
 type Props = {
   evaluateConsole: (command: string) => void,
 };
 
-export default class ConsoleInput extends React.PureComponent<Props> {
+const InputWrapper = styled.div`
+  position: relative;
+  height: 100%;
+  padding-top: ${CONSOLE_INPUT_TOP_PADDING}px;
+  padding-bottom: ${CONSOLE_INPUT_BOTTOM_PADDING}px;
+  box-sizing: border-box;
+  width: 100%;
+`;
+
+const monacoOptions = {
+  // language: 'javascript',
+  wordWrap: 'on',
+  overviewRulerLanes: 0,
+  glyphMargin: false,
+  lineNumbers: 'off',
+  folding: false,
+  selectOnLineNumbers: false,
+  selectionHighlight: false,
+  cursorStyle: 'line-thin',
+  scrollbar: {
+    useShadows: false,
+    horizontal: 'hidden',
+    verticalScrollbarSize: 9,
+  },
+  lineDecorationsWidth: 0,
+  scrollBeyondLastLine: false,
+  renderLineHighlight: 'none',
+  minimap: {
+    enabled: false,
+  },
+  contextmenu: false,
+  ariaLabel: 'ConsoleInput',
+  fontFamily: 'Menlo, monospace',
+  fontSize: 13,
+};
+
+function noop() {
+  return Promise.resolve();
+}
+
+class ConsoleInput extends React.PureComponent<Props> {
+  sizeProbeInterval: number;
+
   state = {
     command: '',
     commandHistory: [],
     commandCursor: -1,
+
+    editorHeight: CONSOLE_INPUT_LINE_HEIGHT,
   };
 
-  onChange = e => {
-    if (e.target) {
-      this.setState({ command: e.target.value });
+  resizeEditor = () => {
+    this.editor.layout();
+  };
+
+  editorDidMount = async (editor, monaco) => {
+    this.editor = editor;
+
+    editor.onDidChangeModelContent(() => {
+      const lineCount = editor.getModel().getLineCount();
+      this.setState({
+        editorHeight: Math.min(
+          CONSOLE_INPUT_MAX_HEIGHT,
+          lineCount * CONSOLE_INPUT_LINE_HEIGHT
+        ),
+      });
+      this.resizeEditor();
+    });
+
+    editor.onKeyDown(event => {
+      const e = event.browserEvent;
+      const { evaluateConsole } = this.props;
+
+      if (e.keyCode === 13) {
+        // Enter
+        if (e.shiftKey) {
+          return;
+        }
+        e.preventDefault();
+        e.stopPropagation();
+        const command = editor.getModel().getValue();
+        evaluateConsole(command);
+        editor.setValue('');
+        this.setState({
+          commandCursor: -1,
+          commandHistory: [command, ...this.state.commandHistory],
+        });
+      } else if (e.keyCode === 38) {
+        // Up arrow
+        const lineNumber = editor.getPosition().lineNumber;
+        if (lineNumber !== 1) {
+          return;
+        }
+
+        const newCursor = Math.min(
+          this.state.commandCursor + 1,
+          this.state.commandHistory.length - 1
+        );
+        editor.setValue(this.state.commandHistory[newCursor] || '');
+        this.setState({
+          commandCursor: newCursor,
+        });
+      } else if (e.keyCode === 40) {
+        // Down arrow
+        const lineNumber = editor.getPosition().lineNumber;
+        const lineCount = editor.getModel().getLineCount();
+        if (lineNumber !== lineCount) {
+          return;
+        }
+
+        const newCursor = Math.max(this.state.commandCursor - 1, -1);
+        editor.setValue(this.state.commandHistory[newCursor] || '');
+        this.setState({
+          commandCursor: newCursor,
+        });
+      }
+    });
+
+    window.addEventListener('resize', this.resizeEditor);
+    this.sizeProbeInterval = setInterval(this.resizeEditor.bind(this), 3000);
+  };
+
+  componentWillUnmount() {
+    window.removeEventListener('resize', this.resizeEditor);
+    clearTimeout(this.sizeProbeInterval);
+    if (this.editor) {
+      this.editor.dispose();
     }
-  };
-
-  handleKeyDown = e => {
-    const { evaluateConsole } = this.props;
-
-    if (e.keyCode === 13) {
-      e.preventDefault();
-      e.stopPropagation();
-      // Enter
-      evaluateConsole(this.state.command);
-      this.setState({
-        command: '',
-        commandHistory: [this.state.command, ...this.state.commandHistory],
-      });
-    } else if (e.keyCode === 38) {
-      const newCursor = Math.min(
-        this.state.commandCursor + 1,
-        this.state.commandHistory.length - 1
-      );
-      // Up arrow
-      this.setState({
-        command: this.state.commandHistory[newCursor] || '',
-        commandCursor: newCursor,
-      });
-    } else if (e.keyCode === 40) {
-      const newCursor = Math.max(this.state.commandCursor - 1, -1);
-      // Down arrow
-      this.setState({
-        command: this.state.commandHistory[newCursor] || '',
-        commandCursor: newCursor,
-      });
-    }
-  };
+  }
 
   render() {
     return (
-      <Container>
+      <Container height={this.state.editorHeight + 12}>
         <IconContainer style={{ color: theme.secondary() }}>
           <ChevronRight />
         </IconContainer>
-        <Input
-          value={this.state.command}
-          onChange={this.onChange}
-          onKeyUp={this.handleKeyDown}
-        />
+        <InputWrapper>
+          <MonacoEditor
+            width="100%"
+            height={`calc(100% - ${CONSOLE_INPUT_PADDING}px)`}
+            options={monacoOptions}
+            editorDidMount={this.editorDidMount}
+            openReference={noop}
+          />
+        </InputWrapper>
       </Container>
     );
   }
 }
+
+export default ConsoleInput;
