+++ packages/app/src/app/components/sandbox/Preview/DevTools/Console/index.js
@@ -7,7 +7,7 @@ import ClearIcon from 'react-icons/lib/md/clear-all';
 
 import CircularJSON from 'circular-json';
 import Message from './Message';
-import Input from './MonacoInput';
+import Input from './Input';
 
 const Container = styled.div`
   background-color: ${props => props.theme.background};
